Update bank_pos.pl
The variables vx and vy seem to be defined wrt to vertex where the particle was created and then entered the first sensitive volume. This would mean they record the vertex when the gamma rays enter the "radiator flux" element. Not updated every time the particle enters a sensitive volume?